fixed types in PID initialization, moved gyro based drive to its own command
the Rotate command is currently a noop and the real code is in the new
DriveTurn.java class
There were a few references to 'r' constants being used to initialize
'left' motors so these were changed to 'l' constants. This doesn't
change any current behavior because the 'l' and 'r' constants had the
same values.