WARNING!!! lots of edits - PLEASE TEST
After reading the wpilib tutorials for command systems I re-read the
code and tried to clean up our implementation to match best practices.
This resulted in a structure where we have ioDevices, subsystems, and
commands. The result was a lot of little edits.

I also normalized some of the code so it is easier to read and to make
classes in each category more similar - eg command classes look alike
and have a similar layout, subsystem classes look alike and have a
similar layout, etc.

I added some TODO and some FIXME comments. The FIXME need to be reviewed
now and the TODO comments can be reviewed after the season.

To review FIXME and TODO tasks, use the eclipse option: Windows -> Show
View -> Tasks